pyproject.toml
[project]
name = "ytb-gateway"
version = "0.1.0"
description = "Gateway service for YouTube navigation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "pydantic>=2.11.4",
    "python-multipart>=0.0.20",
    "redis>=6.1.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "yt-dlp==2025.5.17.232915.dev0",
    "autogen>=0.9.2",
    "brotli>=1.1.0",
    "mutagen>=1.47.0",
    "pycryptodomex>=3.23.0",
    "websockets>=15.0.1",
    "ag2[openai]>=0.9.2",
]
